# appendix_c_openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Agentic Learning Path API",
    "description": "Multi-Agent AI system for personalized learning paths",
    "version": "0.2.0"
  },
  "paths": {
...
    "/api/v1/learners/signup": {
      "post": {
        "tags": [
          "Learners"
        ],
        "summary": "Signup Learner",
        "description": "Register a new learner for the pilot.\nAssign to COHORT (A/B testing) automatically.",
        "operationId": "signup_learner_api_v1_learners_signup_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/LearnerSignupRequest"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/LearnerProfileResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/v1/learners/pre-test": {
      "post": {
        "tags": [
          "Learners"
        ],
        "summary": "Submit Pre Test",
        "description": "Submit pre-test answers. Calculates score and updates profile.",
        "operationId": "submit_pre_test_api_v1_learners_pre_test_post",
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PreTestSubmission"
              }
            }
          },
          "required": true
        },
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "LearnerProfileResponse": {
        "properties": {
          "learner_id": {
            "type": "string",
            "title": "Learner Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "cohort": {
            "type": "string",
            "title": "Cohort"
          },
          "created_at": {
            "type": "string",
            "format": "date-time",
            "title": "Created At"
          }
        },
        "type": "object",
        "required": [
          "learner_id",
          "name",
          "cohort",
          "created_at"
        ],
        "title": "LearnerProfileResponse"
      },
      "LearnerSignupRequest": {
        "properties": {
          "name": {
            "type": "string",
            "title": "Name"
          },
          "email": {
            "type": "string",
            "format": "email",
            "title": "Email"
          },
          "age": {
            "type": "integer",
            "title": "Age"
          },
          "background": {
            "type": "string",
            "title": "Background"
          },
          "consent_agreed": {
            "type": "boolean",
            "title": "Consent Agreed"
          }
        },
        "type": "object",
        "required": [
          "name",
          "email",
          "age",
          "background",
          "consent_agreed"
        ],
        "title": "LearnerSignupRequest"
      },
...
      "PreTestSubmission": {
        "properties": {
          "learner_id": {
            "type": "string",
            "title": "Learner Id"
          },
          "answers": {
            "additionalProperties": true,
            "type": "object",
            "title": "Answers"
          }
        },
        "type": "object",
        "required": [
          "learner_id",
          "answers"
        ],
        "title": "PreTestSubmission"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}